Basic multithread implementation working
With a progress thread. This is ugly but could be improved with a
progress bar class. The current implementation is simple. Farsite5 class
already had a simple member called "progress". This was a double and was
updated by the main time step loop. I changed this to atomic_int to make
it thread-safe and now instead of a proportion it represents an integer
percentage. The CFarsite class that contains a Farsite5 class in turn
provides read access to this member.

I have a simple progress function loop and poll this member in each
separate CFarsite instance. This is hacky and relies on casting a void
pointer as in the original code. Additionally, this would be nicer as
an actual progress bar rather than just newline updates.

In addition to this progress thread, each CFarsite instance is handed to
a separate thread that launches the instance. These threads are stored
in a std::vector. main() is in charge of all memory management so these
threads cannot outlive main().

+++ src/FARSITE.cpp
@@ -148,16 +148,10 @@ int CFarsite::LaunchFarsite(void)
 }
 
 /****************************************************************************/
-double CFarsite::GetFarsiteProgress()//int *ai_RunStatus, char cr[])
+int CFarsite::GetFarsiteProgress()//int *ai_RunStatus, char cr[])
 {
-
-
 	if (m_pFarsite)
     return m_pFarsite->GetFarsiteProgress();//FPC.GetProgress(m_pFarsite, &m_pFarsite->WN2,&m_pFarsite->cfmc, cr);
-
-//  Original
-//	  	return m_pFarsite->GetFarsiteProgress(ai_RunStatus);
-
 	return 0;
 }
